Remove left join for ofec_committee_totals_per_cycle_vw Add test case for candidate committee no finanical
